rgw-admin: Fix bucket list with no --uid
http://tracker.ceph.com/issues/5228 fixes #5228

The command bucket list with no uid specified was no displaying
the right output. Instead of listing all buckets, it was outputing
nothing.

The is_user_op was returning true if the user was anonymous,
and the list_bucket_init function call was not checking for the
right return value.

Signed-off-by: Christophe Courtaut <christophe.courtaut@gmail.com>