Workaround repopriority of yum local repo.
Because of issues with package name differences vps are
setup to use repopriority and our local repo (which has
some ceph/librados stuff in it) gets high priority so
the ceph.repo that is created on the machine from
ceph-release basically gets ignored. This change makes
it so ceph.repo is the same priority level as our local
repo.


Signed-off-by: Sandon Van Ness <sandon@inktank.com>