Revert "mon/OSDMonitor: send_to_waiting() in on_active()"
This reverts commit f06a124a7fa0717ef8c523408b31d814df57caca.

On peons, on_active() is only called when we *first* become active after an
election.  Only on the leader is it called after each commit/update.  This
makes this change cause other problems (broken subscriptions on peons, in
particular).  We possibly should fix that, but there is also a simpler fix
for the original problem we were trying to solve.

Signed-off-by: Sage Weil <sage@inktank.com>